Move CAML/DataBlocks to util package
