+++ src/main/java/cam72cam/immersiverailroading/gui/overlay/GuiBuilder.java
@@ -3,6 +3,7 @@ package cam72cam.immersiverailroading.gui.overlay;
 import cam72cam.immersiverailroading.ImmersiveRailroading;
 import cam72cam.immersiverailroading.entity.EntityRollingStock;
 import cam72cam.immersiverailroading.library.GuiText;
+import cam72cam.immersiverailroading.util.DataBlock;
 import cam72cam.mod.event.ClientEvents;
 import cam72cam.mod.gui.helpers.GUIHelpers;
 import cam72cam.mod.math.Vec3d;
