+++ src/main/java/cam72cam/immersiverailroading/registry/FreightDefinition.java
@@ -1,7 +1,7 @@
 package cam72cam.immersiverailroading.registry;
 
 import cam72cam.immersiverailroading.entity.EntityRollingStock;
-import cam72cam.immersiverailroading.gui.overlay.DataBlock;
+import cam72cam.immersiverailroading.util.DataBlock;
 import cam72cam.immersiverailroading.model.FreightModel;
 import cam72cam.immersiverailroading.model.StockModel;
