+++ src/main/java/cam72cam/immersiverailroading/util/CAML.java
@@ -1,4 +1,4 @@
-package cam72cam.immersiverailroading.gui.overlay;
+package cam72cam.immersiverailroading.util;
 
 import org.apache.commons.io.IOUtils;
 import org.apache.commons.lang3.StringUtils;
